--- src/features/common/Layout/Navbar/utils.ts
@@ -1,2 +1,0 @@
-export const RESTORATION_ORGANIZATIONS_URL =
-  'https://www.plant-for-the-planet.org/restoration-organizations';
